fix(neko_roast): harden live event selection window
